Project Restructuring...
I saw that our previous project structure was all over the place so i restructured it to make it more efficient to use. Everything work on my local device so just let me know if you encounter any issues and i can revert back to our backup if need be.